+++ Samples~/Scripts/Wwise/GetBankExample.cs
@@ -1,3 +1,5 @@
+using System;
+using System.Collections.Generic;
 using AK.Wwise;
 using MaJiang.Utils;
 using SaintsField.Wwise;
@@ -8,8 +10,6 @@ namespace SaintsField.Samples.Scripts.Wwise
 {
     public class GetBankExample : SaintsMonoBehaviour
     {
-        [SerializeField, GetWwise("PlayBGM")] private AK.Wwise.Event _playEvent;
-
         [GetWwise("BGM*")]
         public Bank bank;
 
@@ -23,5 +23,40 @@ namespace SaintsField.Samples.Scripts.Wwise
 
         [GetWwise("*/BGM/Stop*")]
         public Event[] stopEvents;
+
+        [DefaultExpand] public SaintsDictionary<int, Switch> bgmSwitchDictioary;
+
+        [GetWwise("Normal")]
+        public Switch bgSwitch;
+
+        [Serializable]
+        public class BGMSwitchDictioary: SaintsDictionaryBase<int, Switch>
+        {
+            protected override List<Wrap<int>> SerializedKeys => _saintsKeys;
+            protected override List<Wrap<Switch>> SerializedValues => _saintsValues;
+
+#if UNITY_EDITOR
+            // ReSharper disable once UnusedMember.Local
+            private static string EditorPropKeys => nameof(_saintsKeys);
+            // ReSharper disable once UnusedMember.Local
+            private static string EditorPropValues => nameof(_saintsValues);
+#endif
+
+            [SerializeField] private List<Wrap<int>> _saintsKeys = new List<Wrap<int>>();
+
+#if UNITY_EDITOR
+            [GetWwise("$" + nameof(EditorGetSwitch))]
+#endif
+            [SerializeField] private List<Wrap<Switch>> _saintsValues = new List<Wrap<Switch>>();
+
+#if UNITY_EDITOR
+            private string EditorGetSwitch(int index)
+            {
+                return $"//Switch{index}";
+            }
+#endif
+        }
+
+        [DefaultExpand] public BGMSwitchDictioary bgmSwitchDict;
     }
 }